+++ src/PSDocs/Runtime/PSDocs.cs
@@ -1,4 +1,5 @@
 ﻿using System.Management.Automation;
+using System.Threading;
 
 namespace PSDocs.Runtime
 {
@@ -25,6 +26,41 @@ namespace PSDocs.Runtime
             {
                 return GetContext().TargetObject;
             }
+        }
+
+        /// <summary>
+        /// The current culture.
+        /// </summary>
+        public string Culture
+        {
+            get
+            {
+                return GetContext().Culture;
+            }
+        }
+
+
+        //public string Generator
+        //{
+        //    get
+        //    {
+        //        return GetContext().Generator;
+        //    }
+        //}
+
+        /// <summary>
+        /// Format a string with arguments.
+        /// </summary>
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Exposed as instance method for PowerShell.")]
+        public string Format(string value, params object[] args)
+        {
+            if (string.IsNullOrEmpty(value))
+                return string.Empty;
+
+            if (args == null || args.Length == 0)
+                return value;
+            else
+                return string.Format(Thread.CurrentThread.CurrentCulture, value, args);
         }
 
         private RunspaceContext GetContext()